refactor sendImage to use hooks
